docker-requirements.txt
flask
flask_api
flask_mail
requests
pyyaml
lookerapi
psycopg2